Automate Build and Deploy to GitHub Pages
Github pages requires the use of Jekyll, which while powerful is very
much a blog-centric static site assembler.

Further, Jekyll is a _Ruby-centric_ tool.
I personally prefer writing Ruby code, but I prefer even more to keep
the amount of tools and languages someone needs to know to contribute to
a project to as small a number as possible.

Because this project contains JavaScript for other purposes; I am
choosing to lean on JavaScript for assembling and distributing the
website as well.