Make index.js more parseable at a glance
The longer a file is, the more difficult it is to determine what that
particular file does at a glance.

Further, I have a slight preference that any given file either defines a module
for use in other modules, or combines modules to perform a particular
function.

By splitting out build and publish into discrete modules, I encapsulate
responsibility for these actions with responsibility for interacting
with the external modules they use.