Attempt to prevent redirect loop
I do not know why this is so difficult for me, but I am struggling to
figure out how to configure a github page so that both www and the bare
domain show the same website.

Currently I have Cloudflare set up with CNAME flattening for
zinc.coop and www.zinc.coop pointing to zinc-collective.github.io and a
redirect rule directing all zinc.coop traffic to https://www.zinc.coop

I _believe_ this means that the CNAME record in Github must be
www.zinc.coop instead of zinc.coop.