Provide more debugging information
The ghpages library does not appear to be taking the changes when
invoked programmaticaly. I am sure I am doing something horribly wrong;
but am not sure what that is.

My best guess is it has something to do with the fact that I'm passing
in relative paths instead of absolute paths, but don't have time to
diagnose further.