Format the Operating Agreement up to Section C
Before diving into turning this into a web site builder, I wanted to
take a moment to read through it and format it as plain HTML with ID's
for deep-linking.

This serves two purposes:

1. It re-familiarizes me with all the details; so I can better comply
   with the agreements we've made _and_
2. It gives me a chance to think about the things I may want to work on
   for a v1.1 of the agreement (for instance, including a
   consumer-owner component, or adding explicit seams for modularity if
   others want to use it for their own cooperative)

Unfortunately I ran out of time for today. If someone else carries it
forward I wouldn't be sad.