Generate pdf from the operating agreement (#8)
Automates generating a PDF of the operating agreement as part of the build process. I leveraged the [`prebuild` hook built into npm](https://www.marcusoft.net/2015/08/pre-and-post-hooks-for-npm-scripting.html) for now; but I could see adopting `make` if it starts getting unwieldly.

I also added a first draft of how to contribute to the README, because I am sure future Zee will forget how to do it.