📝 Publish default code of conduct
I'm starting with the contributor covenant, not because it is the best code of conduct; but because it's the one I'm most familiar with.

I generally prefer to adapt a code of conduct from an initial starting point.
In the  interest of expedience I'm going to start here and hope we
spend some more thought and time on this.

Also, as we develop a wider surface area of pages; it may make sense to
place some of the less-important links at the bottom of the site?